Update bounce to use scanamo instead of using dynamodb classed directly. this helped to removed a bunch of boiler plate code.
