+++ src/main/scala/io/onema/bounce/BounceFunction.scala
@@ -14,7 +14,7 @@ package io.onema.bounce
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBAsyncClientBuilder
 import com.amazonaws.services.lambda.runtime.Context
 import com.amazonaws.services.lambda.runtime.events.SNSEvent
-import io.onema.bounce.Logic.SesNotification
+import io.onema.bounce.BounceLogic.SesNotification
 import io.onema.json.Extensions._
 import io.onema.userverless.configuration.lambda.EnvLambdaConfiguration
 import io.onema.userverless.function.LambdaHandler
@@ -24,14 +24,13 @@ import scala.collection.JavaConverters._
 class BounceFunction extends LambdaHandler[SNSEvent, Unit] with EnvLambdaConfiguration {
 
   //--- Fields ---
-  val logic = new Logic(
+  val logic = new BounceLogic(
     AmazonDynamoDBAsyncClientBuilder.defaultClient(),
     getValue("/table/name").getOrElse("LambdaMailerSESNotifications")
   )
 
   //--- Methods ---
   override def execute(event: SNSEvent, context: Context): Unit = {
-    log.info(event.asJson)
     val snsRecord: SNSEvent.SNS = event.getRecords.asScala.head.getSNS
     val snsPublishTime = snsRecord.getTimestamp.toString("yyyy-MM-dd'T'HH:mm:ss'Z'")
     val sesMessage = snsRecord.getMessage.jsonDecode[SesNotification]
